fix(projects/products/stable/platform_ide/ui): add dioxus dependency and externalize view tests
Part of #1075
