+++ projects/products/stable/platform_ide/ui/src/diff_view.rs
@@ -24,42 +24,3 @@ impl DiffView {
         self.lines.iter().any(|l| l.kind != DiffLineKind::Context)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn load_sets_fields() {
-        let mut view = DiffView::default();
-        view.load(
-            "src/main.rs".to_string(),
-            vec![
-                DiffLineEntry {
-                    kind: DiffLineKind::Context,
-                    content: "fn main() {}".to_string(),
-                },
-                DiffLineEntry {
-                    kind: DiffLineKind::Added,
-                    content: "println!(\"hello\");".to_string(),
-                },
-            ],
-        );
-        assert_eq!(view.path.as_deref(), Some("src/main.rs"));
-        assert_eq!(view.lines.len(), 2);
-        assert!(view.has_changes());
-    }
-
-    #[test]
-    fn no_changes_when_all_context() {
-        let mut view = DiffView::default();
-        view.load(
-            "a.txt".to_string(),
-            vec![DiffLineEntry {
-                kind: DiffLineKind::Context,
-                content: "unchanged".to_string(),
-            }],
-        );
-        assert!(!view.has_changes());
-    }
-}
