+++ projects/products/stable/platform_ide/ui/src/issue_list_view.rs
@@ -20,23 +20,3 @@ impl IssueListView {
         self.loading = false;
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn set_issues_updates_list() {
-        let mut view = IssueListView {
-            loading: true,
-            ..IssueListView::default()
-        };
-        view.set_issues(vec![IssueEntry {
-            id: "issue-1".to_string(),
-            name: "My Issue".to_string(),
-            description: None,
-        }]);
-        assert_eq!(view.issues.len(), 1);
-        assert!(!view.loading);
-    }
-}
