+++ projects/products/stable/platform_ide/ui/src/tests/mod.rs
@@ -1,0 +1,9 @@
+mod auth_view;
+mod change_submit_view;
+mod diff_view;
+mod editor_view;
+mod issue_list_view;
+mod offline_controls;
+mod slice_explorer;
+mod ui_app;
+mod verification_view;
